backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psutil==7.0.0
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
aiosqlite==0.21.0
python-multipart==0.0.20
